module added, updated ui + minor modifications
