+++ start.html
@@ -21,7 +21,7 @@
           Print them out. Make meeting notes, project notes, to-do lists—with Quire as your
           note taking app.
         </p>
-        <button type="sumbit" class="btn" onclick="location.href = '/';">Create a note</button>
+        <button type="sumbit" class="btn btn-join" onclick="location.href = '/';">Create a note</button>
       </div>
       <img src="/assets/img/guy_desk.svg" alt="guy at desk"></img>
     </div>
